edit comment logic - mvc

+++ js/index.js
@@ -4,37 +4,39 @@ let url = window.location.pathname;
 url = url.split("/");
 
 watchlistBtn.addEventListener("click", (e) => {
-    if(watchlistBtn.firstElementChild.innerText == "favorite") {
-        try {
-            const requestOptions = {
-                method: "DELETE",
-                headers: { "Content-Type": "application/json" },
-                body: JSON.stringify({ movie_id: url[url.length - 1], watchlist: true})
-            };
-            
-            fetch(`/movies`, requestOptions)
-                .then(response => response.json());
-        } catch (error) {
-            console.log(error);
-        }
-    
-        watchlistBtn.firstElementChild.textContent = "favorite_border";
-    } else {
-        try {
-            const requestOptions = {
-                method: "POST",
-                headers: { "Content-Type": "application/json" },
-                body: JSON.stringify({ movie_id: url[url.length - 1], watchlist: true })
-            };
-            
-            fetch(`/movies`, requestOptions)
-                .then(response => response.json());
-        } catch (error) {
-            console.log(error);
-        }
-    
-        watchlistBtn.firstElementChild.textContent = "favorite";
-    };
+    if(document.cookie.indexOf('token=') !== -1) {
+        if(watchlistBtn.firstElementChild.innerText == "favorite") {
+            try {
+                const requestOptions = {
+                    method: "DELETE",
+                    headers: { "Content-Type": "application/json" },
+                    body: JSON.stringify({ movie_id: url[url.length - 1], watchlist: true})
+                };
+                
+                fetch(`/movies`, requestOptions)
+                    .then(response => response.json());
+            } catch (error) {
+                console.log(error);
+            }
+        
+            watchlistBtn.firstElementChild.textContent = "favorite_border";
+        } else {
+            try {
+                const requestOptions = {
+                    method: "POST",
+                    headers: { "Content-Type": "application/json" },
+                    body: JSON.stringify({ movie_id: url[url.length - 1], watchlist: true })
+                };
+                
+                fetch(`/movies`, requestOptions)
+                    .then(response => response.json());
+            } catch (error) {
+                console.log(error);
+            }
+        
+            watchlistBtn.firstElementChild.textContent = "favorite";
+        };
+    }
 });
 
 const reviewsBtns = document.querySelectorAll(".reviews__button");
